fix(ui): split family catalog modules
