+++ npm/ui/core/src/family-catalog-basics.ts
@@ -1,0 +1,163 @@
+import {
+  catalogOwner,
+  componentQualityGates,
+  interactionQualityGates,
+  stableQualityGates,
+  type UiFamilyCatalogEntry,
+} from "./family-catalog-types.ts";
+
+export const basicFamilyCatalog = [
+  {
+    canonicalName: "button",
+    title: "Button",
+    packageSubpath: "./button",
+    entryFile: "src/button.ts",
+    sourceFiles: ["src/ActionButton.vue", "src/button.ts", "src/button-keyboard.ts"],
+    behaviorContract: "src/button.behavior.md",
+    tests: ["src/button.test.ts"],
+    rendererFixture: "ActionButton.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Button",
+      retainedSignature: "aria-busy",
+      maximumJavaScriptGzipBytes: 1_000,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["action-button", "button primitive"],
+    upstreamCoverage: ["shadcn/ui Button", "Reka UI Primitive", "React Aria Button"],
+    dependencies: ["primitive", "press"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "checkbox",
+    title: "Checkbox",
+    packageSubpath: "./checkbox",
+    entryFile: "src/checkbox.ts",
+    sourceFiles: ["src/CheckboxControl.vue", "src/checkbox.ts", "src/checkbox-state.ts"],
+    behaviorContract: "src/checkbox.behavior.md",
+    tests: ["src/checkbox.test.ts"],
+    rendererFixture: "CheckboxControl.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Checkbox",
+      retainedSignature: "aria-checked",
+      allowedRetainedFamilies: ["controllable-state"],
+      maximumJavaScriptGzipBytes: 1_100,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["checkbox control", "mixed checkbox"],
+    upstreamCoverage: ["shadcn/ui Checkbox", "Reka UI Checkbox", "React Aria Checkbox"],
+    dependencies: ["controllable-state"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "collection",
+    title: "Collection Registry",
+    packageSubpath: "./collection",
+    entryFile: "src/collection.ts",
+    sourceFiles: [
+      "src/collection.ts",
+      "src/collection-registry.ts",
+      "src/collection-keys.ts",
+      "src/collection-observer.ts",
+      "src/collection-order.ts",
+      "src/collection-text.ts",
+      "src/collection-types.ts",
+    ],
+    behaviorContract: "src/collection.behavior.md",
+    tests: ["src/collection.test.ts", "src/collection-dom.test.ts"],
+    typeTests: ["src/collection.types.test-d.ts"],
+    qualityGates: stableQualityGates,
+    bundleBudget: {
+      exportName: "createCollectionRegistry",
+      retainedSignature: "VIZE_UI_COLLECTION_DISPOSED",
+      maximumJavaScriptGzipBytes: 3_150,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["collection registry", "item registry", "logical focus registry"],
+    upstreamCoverage: ["React Aria Collection", "Ariakit Composite Store", "Zag collection"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "composite-navigation",
+    title: "Composite Navigation",
+    packageSubpath: "./composite-navigation",
+    entryFile: "src/composite-navigation.ts",
+    sourceFiles: [
+      "src/composite-navigation.ts",
+      "src/composite-navigation-dom.ts",
+      "src/composite-navigation-internal.ts",
+      "src/composite-navigation-types.ts",
+    ],
+    behaviorContract: "src/composite-navigation.behavior.md",
+    tests: [
+      "src/composite-navigation.test.ts",
+      "src/composite-navigation-active-descendant.test.ts",
+      "src/composite-navigation-lifecycle.test.ts",
+      "src/composite-navigation-ssr.test.ts",
+    ],
+    typeTests: ["src/composite-navigation.types.test-d.ts"],
+    rendererFixture: "CompositeNavigationConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createCompositeNavigation",
+      retainedSignature: "VIZE_UI_COMPOSITE_NAVIGATION_DISPOSED",
+      allowedRetainedFamilies: ["typeahead"],
+      maximumJavaScriptGzipBytes: 3_900,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["roving focus", "active descendant", "composite widget navigation"],
+    upstreamCoverage: ["WAI-ARIA composite widgets", "Reka UI Roving Focus", "Ariakit Composite"],
+    dependencies: ["collection", "typeahead"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "context",
+    title: "Typed Component Context",
+    packageSubpath: "./context",
+    entryFile: "src/context.ts",
+    sourceFiles: ["src/context.ts"],
+    behaviorContract: "src/context.behavior.md",
+    tests: ["src/context.test.ts"],
+    typeTests: ["src/context.types.test-d.ts"],
+    qualityGates: stableQualityGates,
+    bundleBudget: {
+      exportName: "createContext",
+      retainedSignature: "VIZE_UI_CONTEXT_MISSING",
+      maximumJavaScriptGzipBytes: 450,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["provide inject context", "compound component context"],
+    upstreamCoverage: ["Reka UI createContext", "Radix context scope"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "controllable-state",
+    title: "Controlled and Uncontrolled State",
+    packageSubpath: "./controllable-state",
+    entryFile: "src/controllable-state.ts",
+    sourceFiles: ["src/controllable-state.ts"],
+    behaviorContract: "src/controllable-state.behavior.md",
+    tests: ["src/controllable-state.test.ts"],
+    typeTests: ["src/controllable-state.types.test-d.ts"],
+    qualityGates: stableQualityGates,
+    bundleBudget: {
+      exportName: "useControllableState",
+      retainedSignature: "defaultValue[\\s\\S]*Object\\.is[\\s\\S]*defaultValue",
+      maximumJavaScriptGzipBytes: 550,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["controllable state", "controlled state", "uncontrolled state"],
+    upstreamCoverage: ["Radix useControllableState", "React Aria controlled state"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+] as const satisfies readonly UiFamilyCatalogEntry[];
