+++ npm/ui/core/src/family-catalog-focus.ts
@@ -1,0 +1,209 @@
+import {
+  catalogOwner,
+  componentQualityGates,
+  interactionQualityGates,
+  type UiFamilyCatalogEntry,
+} from "./family-catalog-types.ts";
+
+export const focusFamilyCatalog = [
+  {
+    canonicalName: "focus",
+    title: "Focus Interactions",
+    packageSubpath: "./focus",
+    entryFile: "src/focus.ts",
+    sourceFiles: ["src/focus.ts", "src/focus-internal.ts", "src/focus-types.ts"],
+    behaviorContract: "src/focus.behavior.md",
+    tests: ["src/focus.test.ts", "src/focus-lifecycle.test.ts", "src/focus-ssr.test.ts"],
+    typeTests: ["src/focus.types.test-d.ts"],
+    rendererFixture: "FocusConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createFocus",
+      retainedSignature: "VIZE_UI_FOCUS_DISPOSED",
+      allowedRetainedFamilies: ["interaction-modality"],
+      maximumJavaScriptGzipBytes: 3_300,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["focus ring", "focus within", "focus visible"],
+    upstreamCoverage: ["React Aria useFocus", "React Aria useFocusRing", "WICG focus-visible"],
+    dependencies: ["interaction-modality"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "focus-guards",
+    title: "Focus Guards",
+    packageSubpath: "./focus-guards",
+    entryFile: "src/focus-guards.ts",
+    sourceFiles: [
+      "src/focus-guards.ts",
+      "src/focus-guards-internal.ts",
+      "src/focus-guards-stack.ts",
+      "src/focus-guards-types.ts",
+    ],
+    behaviorContract: "src/focus-guards.behavior.md",
+    tests: ["src/focus-guards.test.ts", "src/focus-guards-ssr.test.ts"],
+    typeTests: ["src/focus-guards.types.test-d.ts"],
+    rendererFixture: "FocusGuardsConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createFocusGuards",
+      retainedSignature: "VIZE_UI_FOCUS_GUARDS_DISPOSED",
+      maximumJavaScriptGzipBytes: 3_500,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["focus sentinels", "tab guards", "focus trap guards"],
+    upstreamCoverage: ["Radix FocusGuards", "React Aria FocusScope sentinels"],
+    dependencies: ["focus-scope"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "focus-scope",
+    title: "Focus Scope",
+    packageSubpath: "./focus-scope",
+    entryFile: "src/focus-scope.ts",
+    sourceFiles: [
+      "src/focus-scope.ts",
+      "src/focus-scope-dom.ts",
+      "src/focus-scope-internal.ts",
+      "src/focus-scope-manager.ts",
+      "src/focus-scope-stack.ts",
+      "src/focus-scope-types.ts",
+      "src/use-focus-scope.ts",
+    ],
+    behaviorContract: "src/focus-scope.behavior.md",
+    tests: [
+      "src/focus-scope.test.ts",
+      "src/focus-scope-lifecycle.test.ts",
+      "src/focus-scope-ssr.test.ts",
+    ],
+    typeTests: ["src/focus-scope.types.test-d.ts"],
+    rendererFixture: "FocusScopeConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createFocusScope",
+      retainedSignature: "VIZE_UI_FOCUS_SCOPE_DISPOSED",
+      maximumJavaScriptGzipBytes: 4_050,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["focus trap", "focus containment", "focus restoration"],
+    upstreamCoverage: ["React Aria FocusScope", "Radix FocusScope", "Ariakit FocusTrap"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "hover",
+    title: "Hover Interaction",
+    packageSubpath: "./hover",
+    entryFile: "src/hover.ts",
+    sourceFiles: ["src/hover.ts", "src/hover-types.ts"],
+    behaviorContract: "src/hover.behavior.md",
+    tests: ["src/hover.test.ts", "src/hover-ssr.test.ts"],
+    typeTests: ["src/hover.types.test-d.ts"],
+    rendererFixture: "HoverConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createHover",
+      retainedSignature: "VIZE_UI_HOVER_DISPOSED",
+      maximumJavaScriptGzipBytes: 1_450,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["hover", "pointer hover", "pen hover"],
+    upstreamCoverage: ["React Aria useHover", "Pointer Events hover"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "id",
+    title: "Deterministic ID Provider",
+    packageSubpath: "./id",
+    entryFile: "src/id.ts",
+    sourceFiles: ["src/DeterministicIdProvider.vue", "src/id.ts", "src/deterministic-id.ts"],
+    behaviorContract: "src/id.behavior.md",
+    tests: ["src/id.test.ts"],
+    typeTests: ["src/id.types.test-d.ts"],
+    rendererFixture: "DeterministicIdProvider.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "IdProvider",
+      retainedSignature: "DeterministicIdProvider",
+      maximumJavaScriptGzipBytes: 1_050,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["id provider", "deterministic ids", "hydration ids"],
+    upstreamCoverage: ["React Aria SSRProvider", "Vue useId"],
+    dependencies: ["context"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "inert-outside",
+    title: "Outside Inerting",
+    packageSubpath: "./inert-outside",
+    entryFile: "src/inert-outside.ts",
+    sourceFiles: [
+      "src/inert-outside.ts",
+      "src/inert-outside-dom.ts",
+      "src/inert-outside-internal.ts",
+      "src/inert-outside-stack.ts",
+      "src/inert-outside-types.ts",
+    ],
+    behaviorContract: "src/inert-outside.behavior.md",
+    tests: ["src/inert-outside.test.ts", "src/inert-outside-ssr.test.ts"],
+    typeTests: ["src/inert-outside.types.test-d.ts"],
+    rendererFixture: "InertOutsideConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createInertOutside",
+      retainedSignature: "VIZE_UI_INERT_OUTSIDE_DISPOSED",
+      maximumJavaScriptGzipBytes: 2_175,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["aria hidden outside", "outside inerting", "overlay inerting"],
+    upstreamCoverage: ["Radix overlay inerting", "Adobe Spectrum overlay provider"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "interaction-modality",
+    title: "Interaction Modality",
+    packageSubpath: "./interaction-modality",
+    entryFile: "src/interaction-modality.ts",
+    sourceFiles: [
+      "src/interaction-modality.ts",
+      "src/interaction-modality-events.ts",
+      "src/interaction-modality-hub.ts",
+      "src/interaction-modality-types.ts",
+    ],
+    behaviorContract: "src/interaction-modality.behavior.md",
+    tests: [
+      "src/interaction-modality.test.ts",
+      "src/interaction-modality-lifecycle.test.ts",
+      "src/interaction-modality-ssr.test.ts",
+    ],
+    typeTests: ["src/interaction-modality.types.test-d.ts"],
+    qualityGates: [
+      "behavior-contract",
+      "mounted-dom",
+      "type-inference",
+      "ssr",
+      "tree-shaking",
+      "bundle-size",
+    ],
+    bundleBudget: {
+      exportName: "createInteractionModalityTracker",
+      retainedSignature: "VIZE_UI_INTERACTION_MODALITY_DISPOSED",
+      maximumJavaScriptGzipBytes: 1_650,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["focus modality", "input modality", "focus visible modality"],
+    upstreamCoverage: ["React Aria useFocusVisible", "WICG focus-visible"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+] as const satisfies readonly UiFamilyCatalogEntry[];
