+++ npm/ui/core/src/family-catalog-interactions.ts
@@ -1,0 +1,223 @@
+import {
+  catalogOwner,
+  componentQualityGates,
+  interactionQualityGates,
+  type UiFamilyCatalogEntry,
+} from "./family-catalog-types.ts";
+
+export const interactionFamilyCatalog = [
+  {
+    canonicalName: "long-press",
+    title: "Long Press Interaction",
+    packageSubpath: "./long-press",
+    entryFile: "src/long-press.ts",
+    sourceFiles: ["src/long-press.ts", "src/long-press-internal.ts", "src/long-press-types.ts"],
+    behaviorContract: "src/long-press.behavior.md",
+    tests: [
+      "src/long-press.test.ts",
+      "src/long-press-hardening.test.ts",
+      "src/long-press-legacy.test.ts",
+      "src/long-press-ssr.test.ts",
+    ],
+    typeTests: ["src/long-press.types.test-d.ts"],
+    rendererFixture: "LongPressConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createLongPress",
+      retainedSignature: "VIZE_UI_LONG_PRESS_DISPOSED",
+      allowedRetainedFamilies: ["press"],
+      maximumJavaScriptGzipBytes: 5_150,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["long press", "press and hold"],
+    upstreamCoverage: ["React Aria useLongPress"],
+    dependencies: ["press"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "move",
+    title: "Move Interaction",
+    packageSubpath: "./move",
+    entryFile: "src/move.ts",
+    sourceFiles: ["src/move.ts", "src/move-internal.ts", "src/move-types.ts"],
+    behaviorContract: "src/move.behavior.md",
+    tests: ["src/move.test.ts", "src/move-internal.test.ts", "src/move-ssr.test.ts"],
+    typeTests: ["src/move.types.test-d.ts"],
+    rendererFixture: "MoveConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createMove",
+      retainedSignature: "VIZE_UI_MOVE_DISPOSED",
+      maximumJavaScriptGzipBytes: 2_750,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["move", "drag movement", "keyboard move"],
+    upstreamCoverage: ["React Aria useMove", "Pointer Events pointer capture"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "press",
+    title: "Press Interaction",
+    packageSubpath: "./press",
+    entryFile: "src/press.ts",
+    sourceFiles: [
+      "src/press.ts",
+      "src/press-activation-memory.ts",
+      "src/press-event.ts",
+      "src/press-handlers.ts",
+      "src/press-lifecycle.ts",
+      "src/press-types.ts",
+    ],
+    behaviorContract: "src/press.behavior.md",
+    tests: [
+      "src/press.test.ts",
+      "src/press-lifecycle.test.ts",
+      "src/press-legacy.test.ts",
+      "src/press-ssr.test.ts",
+    ],
+    typeTests: ["src/press.types.test-d.ts"],
+    rendererFixture: "PressConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createPress",
+      retainedSignature: "VIZE_UI_PRESS_DISPOSED",
+      maximumJavaScriptGzipBytes: 3_550,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["press", "activation", "pointer press"],
+    upstreamCoverage: ["React Aria usePress", "Pointer Events activation"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "primitive",
+    title: "Polymorphic Primitive",
+    packageSubpath: "./primitive",
+    entryFile: "src/primitive.ts",
+    sourceFiles: ["src/PrimitiveElement.vue", "src/primitive.ts"],
+    behaviorContract: "src/primitive.behavior.md",
+    tests: ["src/primitive.test.ts"],
+    rendererFixture: "PrimitiveElement.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "Primitive",
+      retainedSignature: "data-vize-ui.+primitive",
+      maximumJavaScriptGzipBytes: 500,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["as child", "polymorphic primitive", "slot forwarding primitive"],
+    upstreamCoverage: ["Reka UI Primitive", "Radix Slot", "shadcn/ui Slot"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "scroll-lock",
+    title: "Scroll Lock",
+    packageSubpath: "./scroll-lock",
+    entryFile: "src/scroll-lock.ts",
+    sourceFiles: [
+      "src/scroll-lock.ts",
+      "src/scroll-lock-dom.ts",
+      "src/scroll-lock-internal.ts",
+      "src/scroll-lock-stack.ts",
+      "src/scroll-lock-types.ts",
+    ],
+    behaviorContract: "src/scroll-lock.behavior.md",
+    tests: ["src/scroll-lock.test.ts", "src/scroll-lock-ssr.test.ts"],
+    typeTests: ["src/scroll-lock.types.test-d.ts"],
+    rendererFixture: "ScrollLockConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createScrollLock",
+      retainedSignature: "VIZE_UI_SCROLL_LOCK_DISPOSED",
+      maximumJavaScriptGzipBytes: 2_150,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["body scroll lock", "document scroll lock", "overlay scroll lock"],
+    upstreamCoverage: ["React Aria usePreventScroll", "Radix RemoveScroll"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "spatial-navigation",
+    title: "Spatial Navigation",
+    packageSubpath: "./spatial-navigation",
+    entryFile: "src/spatial-navigation.ts",
+    sourceFiles: [
+      "src/spatial-navigation.ts",
+      "src/spatial-navigation-internal.ts",
+      "src/spatial-navigation-types.ts",
+    ],
+    behaviorContract: "src/spatial-navigation.behavior.md",
+    tests: [
+      "src/spatial-navigation.test.ts",
+      "src/spatial-navigation-lifecycle.test.ts",
+      "src/spatial-navigation-ssr.test.ts",
+    ],
+    typeTests: ["src/spatial-navigation.types.test-d.ts"],
+    rendererFixture: "SpatialNavigationConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createSpatialNavigation",
+      retainedSignature: "VIZE_UI_SPATIAL_NAVIGATION_DISPOSED",
+      maximumJavaScriptGzipBytes: 2_600,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["spatial navigation", "grid navigation", "geometry navigation"],
+    upstreamCoverage: ["CSS Spatial Navigation", "ARIA grid navigation"],
+    dependencies: ["collection"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "typeahead",
+    title: "Collection Typeahead",
+    packageSubpath: "./typeahead",
+    entryFile: "src/typeahead.ts",
+    sourceFiles: ["src/typeahead.ts", "src/typeahead-internal.ts", "src/typeahead-types.ts"],
+    behaviorContract: "src/typeahead.behavior.md",
+    tests: ["src/typeahead.test.ts", "src/typeahead-ssr.test.ts"],
+    typeTests: ["src/typeahead.types.test-d.ts"],
+    rendererFixture: "TypeaheadConsumer.vue",
+    qualityGates: interactionQualityGates,
+    bundleBudget: {
+      exportName: "createTypeahead",
+      retainedSignature: "VIZE_UI_TYPEAHEAD_DISPOSED",
+      maximumJavaScriptGzipBytes: 1_375,
+      maximumCssGzipBytes: 0,
+    },
+    aliases: ["typeahead", "collection search", "keyboard search"],
+    upstreamCoverage: ["React Aria useTypeSelect", "Ariakit composite typeahead"],
+    dependencies: ["collection"],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+  {
+    canonicalName: "visually-hidden",
+    title: "Visually Hidden",
+    packageSubpath: "./visually-hidden",
+    entryFile: "src/visually-hidden.ts",
+    sourceFiles: ["src/VisuallyHidden.vue", "src/visually-hidden.ts"],
+    behaviorContract: "src/visually-hidden.behavior.md",
+    tests: ["src/visually-hidden.test.ts"],
+    rendererFixture: "VisuallyHidden.vue",
+    qualityGates: componentQualityGates,
+    bundleBudget: {
+      exportName: "VisuallyHidden",
+      retainedSignature: "visually-hidden",
+      maximumJavaScriptGzipBytes: 400,
+      maximumCssGzipBytes: 180,
+    },
+    aliases: ["screen reader only", "sr-only", "visually hidden"],
+    upstreamCoverage: ["React Aria VisuallyHidden", "Radix VisuallyHidden"],
+    dependencies: [],
+    maturity: "stable",
+    owner: catalogOwner,
+  },
+] as const satisfies readonly UiFamilyCatalogEntry[];
